external/cv_maker/requirements.txt
google-genai
python-docx
beautifulsoup4
requests
gdown
pypdf
onedrivedownloader
openai
google-cloud-aiplatform
playwright
google-cloud-storage
rich
